Move group_traces_by_session/get_session_id to utils.py, remove _to_litellm_model duplicate
- Moved get_session_id and group_traces_by_session from sampling.py to utils.py
- Removed _to_litellm_model from constants.py (duplicate of convert_mlflow_uri_to_litellm)
- Updated pipeline.py and test imports accordingly
