+++ mlflow/genai/discovery/sampling.py
@@ -10,13 +10,9 @@ from mlflow.genai.discovery.constants import (
     SAMPLE_POOL_MULTIPLIER,
     SAMPLE_RANDOM_SEED,
 )
-from mlflow.tracing.constant import TraceMetadataKey
+from mlflow.genai.discovery.utils import get_session_id
 
 _logger = logging.getLogger(__name__)
-
-
-def get_session_id(trace: Trace) -> str | None:
-    return (trace.info.trace_metadata or {}).get(TraceMetadataKey.TRACE_SESSION)
 
 
 def sample_traces(
@@ -60,28 +56,3 @@ def sample_traces(
         len(group_keys),
     )
     return result
-
-
-def group_traces_by_session(
-    traces: list[Trace],
-) -> dict[str, list[Trace]]:
-    """
-    Group traces by session ID.
-
-    Traces without a session become standalone single-trace "sessions"
-    keyed by their trace_id. Each group is sorted by timestamp_ms.
-
-    Note: mlflow.genai.evaluation.session_utils has a similar function, but it
-    operates on EvalItem objects and drops traces without sessions. This version
-    works on raw Trace objects and keeps sessionless traces as standalone groups,
-    which is required for the discovery pipeline's frequency calculations.
-    """
-    groups: dict[str, list[Trace]] = defaultdict(list)
-    for trace in traces:
-        session_id = get_session_id(trace) or trace.info.trace_id
-        groups[session_id].append(trace)
-
-    for traces_in_group in groups.values():
-        traces_in_group.sort(key=lambda trace: trace.info.timestamp_ms)
-
-    return dict(groups)
