+++ mlflow/genai/discovery/utils.py
@@ -2,13 +2,46 @@ from __future__ import annotations
 
 import logging
 import threading
+from collections import defaultdict
 
 import mlflow
-from mlflow.genai.discovery.constants import LLM_MAX_TOKENS, NUM_RETRIES, _to_litellm_model
+from mlflow.entities.trace import Trace
+from mlflow.genai.discovery.constants import LLM_MAX_TOKENS, NUM_RETRIES
 from mlflow.genai.discovery.entities import Issue
 from mlflow.genai.judges.adapters.litellm_adapter import _invoke_litellm
+from mlflow.metrics.genai.model_utils import convert_mlflow_uri_to_litellm
+from mlflow.tracing.constant import TraceMetadataKey
 
 _logger = logging.getLogger(__name__)
+
+
+def get_session_id(trace: Trace) -> str | None:
+    return (trace.info.trace_metadata or {}).get(TraceMetadataKey.TRACE_SESSION)
+
+
+def group_traces_by_session(
+    traces: list[Trace],
+) -> dict[str, list[Trace]]:
+    """
+    Group traces by session ID.
+
+    Traces without a session become standalone single-trace "sessions"
+    keyed by their trace_id. Each group is sorted by timestamp_ms.
+
+    Note: mlflow.genai.evaluation.session_utils has a similar function, but it
+    operates on EvalItem objects and drops traces without sessions. This version
+    works on raw Trace objects and keeps sessionless traces as standalone groups,
+    which is required for the discovery pipeline's frequency calculations.
+    """
+    groups: dict[str, list[Trace]] = defaultdict(list)
+    for trace in traces:
+        session_id = get_session_id(trace) or trace.info.trace_id
+        groups[session_id].append(trace)
+
+    for traces_in_group in groups.values():
+        traces_in_group.sort(key=lambda trace: trace.info.timestamp_ms)
+
+    return dict(groups)
 
 
 class _TokenCounter:
@@ -50,7 +83,7 @@ def _call_llm(
     token_counter: _TokenCounter | None = None,
 ) -> object:
     response = _invoke_litellm(
-        litellm_model=_to_litellm_model(model),
+        litellm_model=convert_mlflow_uri_to_litellm(model),
         messages=messages,
         tools=[],
         num_retries=NUM_RETRIES,
